+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 1.0.13
 
 * Filter out players that have a party from the invite dropdown
+* Party Chunk Amount Override is now ignored if the party has the same amount of chunks as the config so when the config
+  changes it will use the config value
 
 # 1.0.12